- delete endpoint

+++ node-app/src/api/recipes/controller.js
@@ -97,8 +97,10 @@ async function updateRecipe(body, params, user, imagePath, thumbnailPath) {
     // eslint-disable-next-line max-len
     if (recipeInDB.user_id.toString() !== user._id.toString()) return { error: errors.RECIPE_UNAUT_USER };
 
-    fs.unlinkSync(recipeInDB.photo_url.imagePath);
-    fs.unlinkSync(recipeInDB.photo_url.thumbnailPath);
+    if (recipeInDB?.photo_url?.imagePath) {
+      fs.unlinkSync(recipeInDB.photo_url.imagePath);
+      fs.unlinkSync(recipeInDB.photo_url.thumbnailPath);
+    }
 
     const photo_url = {
       imagePath,
@@ -113,6 +115,32 @@ async function updateRecipe(body, params, user, imagePath, thumbnailPath) {
     return ({ data: 'Success' });
   } catch (error) {
     fLogger.warn('Unmapped error at updateRecipe', { error });
+    return { error: errors.GENERAL_RECIPE_ERROR };
+  }
+}
+
+async function deleteRecipe(params, user) {
+  const fLogger = logger.child({ function: 'deleteRecipe' });
+  try {
+    const { recipeId } = params;
+    const recipeInDB = await recipeRepository.getRecipeById(recipeId);
+
+    if (!recipeInDB) return { error: errors.RECIPE_FINDONE_BY_ID };
+    // eslint-disable-next-line max-len
+    if (recipeInDB.user_id.toString() !== user._id.toString()) return { error: errors.RECIPE_UNAUT_USER };
+
+    if (recipeInDB?.photo_url?.imagePath) {
+      fs.unlinkSync(recipeInDB.photo_url.imagePath);
+      fs.unlinkSync(recipeInDB.photo_url.thumbnailPath);
+    }
+
+    const result = await recipeRepository.deleteRecipe(recipeId);
+
+    if (!result || result?.deletedCount !== 1) return { error: errors.RECIPE_DELETE };
+
+    return ({ data: 'Success' });
+  } catch (error) {
+    fLogger.warn('Unmapped error at deleteRecipe', { error });
     return { error: errors.GENERAL_RECIPE_ERROR };
   }
 }
@@ -174,4 +202,5 @@ module.exports = {
   searchSuggestions,
   getCategoryRecipes,
   getTagRecipes,
+  deleteRecipe,
 };
